added back the Tests folder in EarlGrey or else we have build problems added clean-build script to delete EarlGrey dependencies to help troubleshoot build issues (resolves #16) added support for multiple schemes (debug/release) (resolves #13)

+++ detox/scripts/test-update-native.js
@@ -20,6 +20,13 @@ if (shell.exec('react-native run-ios').code !== 0) {
 }
 
 shell.echo('\n#################################################################');
+shell.echo('# react-native run-ios --scheme "example Release"');
+if (shell.exec('react-native run-ios --scheme "example Release"').code !== 0) {
+  shell.echo('error: react-native run-ios --scheme "example Release"');
+  process.exit(1);
+}
+
+shell.echo('\n#################################################################');
 shell.echo('# killing ios simulator');
 shell.exec('killall "Simulator"');
